Struct constructor/destructor refactor
